pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "electrofacies"
version = "1.0.0"
description = "Production-quality electrofacies prediction system for the Delaware Mountain Group"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Ander Martinez-Donate"},
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "xgboost>=1.7",
    "imbalanced-learn>=0.11",
    "lasio>=0.31",
    "matplotlib>=3.7",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "click>=8.1",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
app = [
    "streamlit>=1.30",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "shap>=0.42",
]
shap = [
    "shap>=0.42",
]

[project.scripts]
electrofacies = "electrofacies.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120
target-version = "py39"
